Move cirq.ion.ion_gates.MSGate to cirq.ops module
