+++ cirq-core/cirq/ion/__init__.py
@@ -14,7 +14,7 @@
 
 """Trapped ion devices, gates, and compiling utilties."""
 
-from cirq.ion.ion_gates import ms
+from cirq.ops import ms
 
 from cirq.ion.ion_decomposition import two_qubit_matrix_to_ion_operations
